Fixed compilation failure due to scala-ide/scala-ide@c68b3b7d926324b02fc5826029a9741eb4d2bbc0
In the above mentioned commit, the signature of
`InteractiveCompilationUnitEditor.getInteractiveCompilationUnit` was changed.

The only place where I was not entirely sure the refactoring is safe is in
``TemplateConfiguration.getTextHover`` However, I believe the change is ok
because the ``TemplateConfiguration`` can only be used by a ``TemplateEditor``,
hence there is no need to check for the ``contentType`` (and hence the
conditional expression is useless).