Graceful handling of parsing errors from template compiler
* Because the Play template compiler throws an exception as soon as a parsing error is
encountered, calls to the Play template compiler are now wrapped in a ``Try``.
It's the responsability of the calling code to handle the failure case as it
makes sense.

* Use ``Annotation`` for reporting both parsing and compilation errors from
template files. The implication of this is that compilation  errors are no longer
displayed in the Problem View or the Package Explorer.

* Because we now solely use Annotation, we no longer need to use the Eclipse
extension point for defining new markers (look at the deleted lines in
plugin.xml).

* Finally, I'm aware that the code could use some more clean-up, but my intention
was to minimize changes as we plan to release 0.1 possibly next week already.

Fixes #29