Automatically add Scala nature when opening Play2 Template editor
Java Play2 Template project don't have the Scala nature by default, and this
caused an exception to be thrown whenever the user tried to open a Play2
Template editor because the editor needs access to the ``ScalaProject``
facilities.

The implemented solution is to automatically add the Scala nature if a project
doesn't have it.

Note that ``TemplateCompilationUnit.scalaProject`` is no longer ``lazy``
because I want the code to fail fast if a ``ScalaProject`` cannot be created.

Fix #33