Corrected silly bug in WordFinder
The fact that I had a bug pushed me into writing some unit testing for the WordFinder component
and moving it in the editor bundle, as I believe it can be generally useful. As a consequence of
the above, I created a sdt.editor.tests bundle to hosts the tests for the editor.