Fixed some issue with the partitioning of Route files
* Wrote a number of tests for ROUTE_HTTP partitioning and
fixed a number of bugs related to incorrect partitioning.

* Moved down some logic from ``PlayDocumentPartitioner`` to
``TemplateDocumentPartitioner``. The ``PlayDocumentPartitioner`` is used by
both the template and the route editor to partition the file. The former
implementation of ``PlayDocumentPartitioner.getPartition`` was preventing HTTP
method completions to properly work in some corner case. Hence, I moved the
blamed portion of the code in the ``TemplateDocumentPartitioner`` subclass.
Though, I'm actually not convinced that code is needed in the first place.

Note about the change in ``PlayDocumentPartitioner``

The change in ``PlayDocumentPartitioner`` is a side-effect of the changes I've
made in the ``RoutePartitionTokeniser``. My gut feeling is that the current way
we compute the dirty region should be improved, but I'd rather do this in a
separate PR.

Also, this seemingly innocuous change could lead to new issues in the Template editor,
because the PlayDocumentPartitioner is used by both the Route and the Template editors.
I've done some minimal monkey testing in the template editor after this change, and I haven't
seen any issue, but I can't state I'm confident nothing bad will happen out of this.

Fixes #100