Code completion on a route editor's action now works on a single word.
The problem lied in the MembersLocator subclasses; when creating a MembersComputer they did not handle the case when the prefix was the EmptyPackage. The fix was to handle that case and return the appropriate MembersComputer whose allMembers field was overridden to be the declarations found in the RootClass in addition to the declarations of the prefix itself (i.e. the EmptyPackage).

It should be noted that completion does not work on members of modules or classes found in the default package. However, this should be a separate issue ticket, and should be investigated further to see if play supports action methods coming from these sources before implementing a fix.

Fix #126