Added auto format on save in route editor
Added an executionEventListener that listens to events, checks the event is a save event, checks that the event originated from the correct route editor instance, checks that the auto format route formatter preference is enabled, and then, if all checks are as expected, formats the route file.

Corresponding preference field was added to the Route Editor's formatter preference page.

Fix #117