Correctly display java method parameter names in route action completions
The problem was that when generating parameter names, there was no logic to handle when the symbol is from java code. This fix was to add this logic and and to go fetch a corresponding IMethod from JDT and use it to get the proper name(s) of the parameter(s).

CAVEAT: There all ready existed a passing test for auto completing a java method with parameters that was using the correct/desired parameter names). According to the bug that this commit fixes, that pass should have been failing. We hypothesised that what was happening was that there was no byte code for the java files when running the tests, so the tests were generating java symbol information from the java source files instead (thus giving the proper parameter names). Thus, we tried adding a project build and presentation compiler reset in the test so that the java byte code would be used instead, but that did not reproduce the original bug. Thus, the tests as is don't currently test exactly what we want, and there is no known solution (among Mirco and myself) on how to reproduce this issue in the test suite (however, the issue is easy to produce in an actual eclipse project instance, and I have confirmed that this commit does fix the reported bug there).

Fix #140