URI renaming of invalid/local paths in route editor
There was an off-by-one when try rename paths that didn't start with a leading '/'. In actuality, these types of paths are not accepted by play, which is why they were not handled correctly. To fix I refactored the RouteUri class to account for invalid URI paths, and performed the baseOffset/baseLength calculation per URI in the LocalRename handler.

Also moved RouteUri object methods into companion class

Fix #137