HTTP method completion work's at beginning of line
The issue beforehand was that the tokenizer was inserting default tokens to fill any non-contiguous areas. However, this would cause problems when there existed zero-length http method tokens (the default token would precede the method token, and "overlap" with it, thus taking precedence). The fix was to remove the manually inserted default tokens, and instead return the a default token that fills the contiguous area from the getPartition() method of the PlayDocumentPartitioner if a token at the requested caret offset can not be found.

Fix #129