"Add Route Entry" quick fix now obeys auto formatting
When inserting the route into the route editor, the new line will obey the auto formatting rules (essentially choosing the two max column widths). This functionality was reused from RouteFormattingStrategy, and as a side-effect there is a small refactoring of that file/class to make the needed functionality accessible.

This actually fixes #138 in a different manner than suggested in the ticket. It is arguably "smarter" because it uses the route formatter instead of simply inserting the preference column width.

Fix #138