Fixed a lot of parsing bugs
A large part of the parsing bugs were due simply to the fact that the parser wasn't being given enough context to figure out what was not html/default code. So, instead of trying to being smart for efficiency sake, the whole document is reparsed each time now (instead of the dirty subregions, as before).