Fixed bugs causing annotations in right ruler to not disappear.
The fix was to simply not use SSE's IReporter to report messages. It was providing absolutely no functionality (that I could find/see) that adding markers didn't do, except for the bug!

Also wrapped the marker creation code in a try-finally block to insure that we release the model no matter what.