Removed unneeded code in the ModelParser
The code that's being removed was incorrect to begin with, and thus wasn't having any effect. The editor behaves as expected without it. I've found that the implementation of all the DOM stuff is used for DOM (aka HTML) validation. Thus, when adding HTML validation support is when this part of the implementation will need to be made specialised to be scala template syntax aware.