Fixed weird behaviour in new template editor.
What was essentially happening is that there were text regions that had overlapping bounds. They had overlapping bounds because (a) I was forgetting to change both the region's "length" and the region's "textLength" and (b) AttributeNameRegion (from wst) implemented its adjustTextLength method to always increment its text length by 1 (instead of respecting the passed in parameter). The fix was to reimplement (correctly) AttributeNameRegion in a subclass of AttributeNameRegion (a near full reimplementation had to be made because the fields of the AttributeNameRegion are private, rather than protected..).

Fixs #164