TemplateRegionParser properly handles when scala code overlaps 2+ doc regions.
Certain characters (such as '&' and '<') when part of scala code would cause the XML parser to become "confused" and generated "UNDEFINED" document regions that appear in the middle of the scala code. This would in turn cause the code to not be properly highlighted since the TemplateRegionParser specifically wouldn't handle that scenario.. Now it does!

Fixes #167