New Template Editor now properly highlights @ and { of "reusable code blocks"
A leading { or @ in t "reusable code blocks" weren't properly syntax highlighted because the TemplateRegionParser wasn't looking for the magic at and template braces in default partitions that have the "=" character in it. This fix does a pass over those tokens and splits them into two distinct tokens (one with the '=' and the other with either the '@' or the '{').

Fixes #168