Play 2.4 support for action completion/hyperlinking in routes file
Play 2.4 supports Dependency Injection (DI) -
https://www.playframework.com/documentation/2.4.x/Migration24#Dependency-Injection.
A consequence is that routes can now point to action's methods in either a
module or a class. To know if DI is used, we would have to check the project's
sbt build, and look at the value of the `routesGenerator` setting. If DI is
used, then action's methods in classes should be returned (together with
action's methods in modules, if present).

Unfortunately, it's not possible to check the value of `routesGenerator` in the
build, without having first-class sbt support in Eclipse. Because there is no
way to know whether a Play 2.4 project is using DI, the implemented solution
assumes that it may always be used. This is why both modules and classes members
are returned.