twirl library update in play2 plugin
Because Play2.4 is using Twirl version 1.1.1 lib to process templates
in contrary to Play2.3 which uses Twirl version 1.0.4 there is a need
to deliver Twirl functionalities as plugins to Scala IDE Play2 Plugin.
So now it is possible to choose a Play version in Play Plugin preference
page and then correct Twirl delivers appropriate implementation of
template parser and compiler.

Implementation notes:
 - there is one global setting for Play version. So if user wants to
   work on projects of more than one Play version it is recommended to
   do it in different workspaces. The bottleneck is Twirl parser to
   which is quite hard to pass the project settings (in future cosider
   thread local?)
 - every Twirl version has its counterpart template plugin. Code is
   duplicated so far and it is result of the lack of abstraction
   for Twirl's main functionalities. Maybe in future we would need
   to improve it.

Usage notes:
 - code generated by `TwirlCompiler.compileVirtual(...)` method is
   different from generated by `TwirlCompiler.compile(...)`. The
   main diff is the use of `Html` in place of `HtmlFormat.Appendable`.