Play plugin configured per project
This change allows for setting a Play version (and underlying Twirl
version) per project. The main problem lays in passing the project
settings to Template Parser. Now when `TemplateStructureEditor` is
initialized by `IEditorInput` this `IEditorInput` is used to find
project settings and sets it into `ThreadLocal`. Then when
`IStructuredDocument` is created by `TemplateDocumentLoader` this
setting is retrieved from thread local and set in custom implementation
of `IStructuredDocument`. Finally `IStructuredDocument` sets it in
thread local when any public action is executed on it. It causes that
if `TemplateProcessingProvider` does not find the template version it
tries to find it in thread local.